refactor/#30: animal enum 파일이름 수정
