+++ src/common/enum/animal.enum.ts
