+++ src/users/dtos/get-users-by-animal.dto.ts
@@ -1,6 +1,6 @@
 import { IsEnum, IsNumber, IsOptional, IsString } from "class-validator";
 import { Type } from "class-transformer";
-import { Animal } from "../../common/enum/animal-enum";
+import { Animal } from "../../common/enum/animal.enum";
 import { ApiProperty } from "@nestjs/swagger";
 import { orderByField } from "../enum/orderByFeild-enum";
